feat: append size info to shape
