+++ src/plots/matrix/layer.ts
@@ -5,88 +5,10 @@ import { registerPlotType } from '../../base/global';
 import { LayerConfig } from '../../base/layer';
 import ViewLayer, { ViewConfig } from '../../base/view-layer';
 import { getComponent } from '../../components/factory';
-import { registerShape } from '@antv/g2';
 import MatrixLegend, { MatrixLegendConfig } from './component/legend';
+import { getRectPath, getCirclePath, getCircleCurve } from './shape';
 import './component/label';
 import './component/legend';
-
-function getRectPath(cx, cy, width, height, size) {
-  const w = width * size;
-  const h = height * size;
-  const path = [
-    ['M', cx - w / 2, cy + h / 2],
-    ['Q', cx - w / 2, cy, cx - w / 2, cy - h / 2],
-    ['Q', cx, cy - h / 2, cx + w / 2, cy - h / 2],
-    ['Q', cx + w / 2, cy, cx + w / 2, cy + h / 2],
-    ['Q', cx, cy + h / 2, cx - w / 2, cy + h / 2],
-    ['Z'],
-  ];
-  return path;
-}
-
-function getCirclePath(x, y, size) {
-  const path = [
-    ['M', x, y],
-    ['m', -size, 0],
-    ['a', size, size, 0, 1, 0, size * 2, 0],
-    ['a', size, size, 0, 1, 0, -(size * 2), 0],
-    ['Z'],
-  ];
-  return path;
-}
-
-function getCircleCurve(x, y, size) {
-  // 计算四个角和中点
-  const path = [
-    ['M', x - size, y],
-    ['Q', x - size, y - size, x, y - size],
-    ['Q', x + size, y - size, x + size, y],
-    ['Q', x + size, y + size, x, y + size],
-    ['Q', x - size, y + size, x - size, y],
-    ['Z'],
-  ];
-  return path;
-}
-
-registerShape('polygon', 'rect', {
-  draw(cfg, container) {
-    const points = this.parsePoints(cfg.points);
-    const width = points[2].x - points[0].x;
-    const height = points[0].y - points[1].y;
-    const centerX = points[0].x + width / 2;
-    const centerY = points[1].y + height / 2;
-    /*
-    const path = [
-      ['M', centerX - w / 2, centerY + h / 2],
-      ['L', centerX - w / 2, centerY - h / 2],
-      ['L', centerX + w / 2, centerY - h / 2],
-      ['L', centerX + w / 2, centerY + h / 2],
-      ['Z'],
-    ];
-    */
-    const path = getRectPath(centerX, centerY, width, height, cfg.origin.size);
-    return container.addShape('path', {
-      attrs: {
-        path,
-        fill: cfg.color,
-        opacity: 1,
-      },
-    });
-  },
-});
-
-registerShape('point', 'curvePoint', {
-  draw(cfg, container) {
-    const path = getCirclePath(cfg.x, cfg.y, cfg.size);
-    return container.addShape('path', {
-      attrs: {
-        path,
-        fill: cfg.color,
-        opacity: 1,
-      },
-    });
-  },
-});
 
 export interface MatrixViewConfig extends ViewConfig {
   forceSquare?: boolean;
@@ -417,6 +339,7 @@ export default class MatrixLayer<T extends MatrixLayerConfig = MatrixLayerConfig
     _.each(shapes, (shape) => {
       const data = shape.get('origin')._origin;
       const ratio = 0.3 + scale.scale(data[field]) * 0.6;
+      shape.get('origin').size = ratio;
       const bbox = shape.getBBox();
       const width = bbox.width;
       const height = bbox.height;
@@ -477,6 +400,7 @@ export default class MatrixLayer<T extends MatrixLayerConfig = MatrixLayerConfig
       const centerX = bbox.minX + width / 2;
       const centerY = bbox.minY + height / 2;
       const path = getRectPath(centerX, centerY, this.gridSize[0], this.gridSize[1], 1);
+      shape.get('origin').size = 1;
       shape.stopAnimate();
       shape.animate(
         {
